refactor: use webpack plugin emit client manifest when ssr
